+++ spec/internal/app/validators/custom_each_validator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class CustomEachValidator < ActiveModel::EachValidator
+  def validate_each(_record, _attribute, _value)
+    true
+  end
+end